+++ packages/contracts/source/contracts/p2p/api-node-verifier.ts
@@ -1,4 +1,4 @@
-import { ApiNode } from "./peer";
+import { ApiNode } from "./api-node";
 
 export interface ApiNodeVerifier {
 	verify(apiNode: ApiNode): Promise<boolean>;
