+++ packages/contracts/source/contracts/p2p/enums.ts
@@ -1,0 +1,4 @@
+export enum PeerProtocol {
+	Http = 0,
+	Https = 1,
+}
