+++ packages/contracts/source/contracts/p2p/index.ts
@@ -1,15 +1,17 @@
+export * from "./api-node";
+export * from "./api-node-discoverer";
+export * from "./api-node-processor";
+export * from "./api-node-repository";
+export * from "./api-node-verifier";
 export * from "./broadcaster";
 export * from "./downloader";
 export * from "./endpoints";
+export * from "./enums";
 export * from "./header";
 export * from "./header-service";
 export * from "./log";
 export * from "./nes";
 export * from "./peer";
-export * from "./api-node-discoverer";
-export * from "./api-node-processor";
-export * from "./api-node-repository";
-export * from "./api-node-verifier";
 export * from "./peer-communicator";
 export * from "./peer-connector";
 export * from "./peer-discoverer";
