+++ packages/p2p/source/peer.ts
@@ -28,7 +28,7 @@ export class Peer implements Contracts.P2P.Peer {
 
 	public plugins: Contracts.P2P.PeerPlugins = {};
 
-	public apiNodes: Contracts.P2P.ApiNodes = [];
+	public apiNodes: Contracts.P2P.ApiNode[] = [];
 
 	#header: Contracts.P2P.HeaderData | undefined;
